Export helper functions to their own file
Also remove '*' import from EmailGUI
